NeuralAccelerator/requirements.txt
certifi==2020.12.5
chardet==4.0.0
console-progressbar==1.1.2
idna==2.10
jsonpatch==1.28
jsonpointer==2.0
numpy==1.19.4
Pillow==8.0.1
pyzmq==20.0.0
requests==2.25.1
scipy==1.5.4
six==1.15.0
torch==1.7.1
torchfile==0.1.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0